feat: Initial Network module (#2)
* feat(URL): add property PortOrSpecialSchemePortOr80

* fix(SpecialScheme): change datatype of Port from int to ushort

* feat: initial work on Network module

* fix(RequestService): remove chunk count

* fix(HttpRequestHeader): typo in User-Agent

* fix(HttpRequest): add colon delimiter

* feat(HttpRequest): create functions to build requests from URLs

+++ src/Danaus/Url/SpecialScheme.cs
@@ -6,7 +6,7 @@ public class SpecialScheme
 
     public string Scheme { get; }
 
-    public int? Port { get; }
+    public ushort? Port { get; }
 
     public static SpecialScheme Ftp
     {
@@ -38,7 +38,7 @@ public class SpecialScheme
         get { return new SpecialScheme("ws", 443); }
     }
 
-    private SpecialScheme(string scheme, int? port)
+    private SpecialScheme(string scheme, ushort? port)
     {
         Scheme = scheme;
         Port = port;
